Visual/requirements.txt
streamlit
google-cloud-bigquery
pandas>=1.3.0
db-dtypes
plotly